moved file routes.js to routes folder
