+++ index.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import mongoose from 'mongoose';
 import bodyParser from 'body-parser';
 
-import routes from './routes';
+import routes from './routes/routes';
 
 const app = express();
